Added the ability to resize objects.

+++ src/main/java/components/CanvasObject.java
@@ -4,7 +4,7 @@ import java.awt.*;
 
 public class CanvasObject {
 
-    public int x, y, width, height;
+    public double x, y, width, height;
     public double angle = 0;
     public Color color = Color.RED;
     public boolean selected = false;
